Focus search input field automatically and allow typing right away
This is a partial fix for #9. One can already type into the field when the index gets parsed, though browser will propably not show the input as the parsing blocks rendering. The text will show up in the input once the parsing is done.
It may be worth implementing async parsing.